requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
lxml==4.6.1
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
phue==1.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
rgbxy==0.5
scipy==1.5.4
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.51.0
urllib3==1.25.11
webdriver-manager==3.2.2
webdrivermanager==0.9.0
